day 4 solved, search for xmas
Tricky part was how to effectively troubleshoot.
I was missing a single hit on their smaller example input. Helped when i
made a troubelshooting copy of the dataset to track which letters where
used in a hit, then print that out to check against their example hits.

I had missed to account for when indicies where less than OR EQUAL to 0,
meaning i could never get hits for XMAS when the final letter was on i
or j == 0